stock_trader/requirements.txt
asyncpg
ib_async
pydantic
PyYAML
pytest
requests
